Base Code
Importing code that was made late 2013 from September through December.
Slight tweaks in March 2014 to change the default orientation when run
on Android. Game is playable, but horribly broken. Movement is
pointness, character only attack themselves, and the UI is very cramped
on a phone.

Also included are all of the documents I created while in the process of
originally writing this, such as concept art and weekly logs.